fix: re-request known streams on reconnect
Re-requests plasma streams from the client when reconnecting to the
server. Adds the :event and :args to the state's stream maps so they can
be re-requested later.

Maybe there's a stronger opinion on whether to call on-open,
on-reconnect, and the order of things. For now, this is enough to have
apps work without refresh across server restarts.